Added auth for incomes

+++ controllers/expenseController.js
@@ -43,13 +43,19 @@ const getExpense = async (req, res) => {
 
 const deleteExpense = async (req, res) => {
     const { id } = req.params;
-    Expense.findByIdAndDelete(id)
-        .then((income) => {
-            res.status(200).json({ message: "Expense Deleted" });
-        })
-        .catch((err) => {
-            res.status(500).json({ message: "Server Error" });
-        });
+    expense = await Expense.findById(id);
+    console.log(expense);
+    if (!expense) {
+        return res.status(404).json({ message: "Expense not found" });
+    } else {
+        Expense.findByIdAndDelete(id)
+            .then((income) => {
+                res.status(200).json({ message: "Expense Deleted" });
+            })
+            .catch((err) => {
+                res.status(500).json({ message: "Server Error" });
+            });
+    }
 };
 
 module.exports = { addExpense, getExpense, deleteExpense };
