+++ controllers/incomeController.js
@@ -1,5 +1,18 @@
+const jwt = require("jsonwebtoken");
+const User = require("../models/userModel");
 const Income = require("../models/IncomeModel");
-const addIncome = async (req, res) => {
+const asyncHandler = require("express-async-handler");
+const addIncome = asyncHandler(async (req, res) => {
+    const token = req.headers["x-access-token"] || req.headers["authorization"];
+    if (!token) {
+        res.status(401).json({ message: "Not authorized." });
+    }
+    const bearerToken = token.split(" ")[1];
+    const decoded = jwt.verify(bearerToken, process.env.SECRET_KEY);
+    const user = await User.findById(decoded.id);
+    if (!user) {
+        res.status(401).json({ message: "Not authorized." });
+    }
     const { title, amount, date, category, description } = req.body;
     const income = Income({
         title,
@@ -7,6 +20,7 @@ const addIncome = async (req, res) => {
         date,
         category,
         description,
+        user,
     });
     try {
         //validations
@@ -21,28 +35,61 @@ const addIncome = async (req, res) => {
                 .json({ message: "Amount must be a positive number!" });
         }
         await income.save();
+        console.log(income);
         res.status(200).json({ message: "Income Added" });
     } catch (error) {
         res.status(500).json({ message: "Server Error" });
     }
-};
-const getIncome = async (req, res) => {
+});
+const getIncome = asyncHandler(async (req, res) => {
+    const token = req.headers["x-access-token"] || req.headers["authorization"];
+    if (!token) {
+        res.status(401).json({ message: "Not authorized." });
+    }
+    const bearerToken = token.split(" ")[1];
+    const decoded = jwt.verify(bearerToken, process.env.SECRET_KEY);
+    const user = await User.findById(decoded.id);
+    if (!user) {
+        res.status(401).json({ message: "Not authorized." });
+    }
     try {
-        const incomes = await Income.find().sort({ createdAt: -1 });
+        const incomes = await Income.find({ user: user._id }).sort({
+            createdAt: -1,
+        });
         res.status(200).json(incomes);
     } catch (error) {
         console.log(error);
         res.status(500).json({ message: "Server Error" });
     }
-};
-const deleteIncome = async (req, res) => {
+});
+const deleteIncome = asyncHandler(async (req, res) => {
+    const token = req.headers["x-access-token"] || req.headers["authorization"];
+    if (!token) {
+        res.status(401).json({ message: "Not authorized." });
+    }
+    const bearerToken = token.split(" ")[1];
+    const decoded = jwt.verify(bearerToken, process.env.SECRET_KEY);
+    const user = await User.findById(decoded.id);
+    if (!user) {
+        res.status(401).json({ message: "Not authorized." });
+    }
     const { id } = req.params;
-    Income.findByIdAndDelete(id)
-        .then((income) => {
-            res.status(200).json({ message: "Income Deleted" });
-        })
-        .catch((err) => {
-            res.status(500).json({ message: "Server Error" });
-        });
+    income = await Income.find({ _id: id, user: user._id });
+    console.log(income);
+    if (!income) {
+        return res.status(404).json({ message: "Income not found" });
+    } else {
+        Income.findByIdAndDelete(id)
+            .then((income) => {
+                res.status(200).json({ message: "Income Deleted" });
+            })
+            .catch((err) => {
+                res.status(500).json({ message: "Server Error" });
+            });
+    }
+});
+const verifyToken = (token) => {
+    const { id, username } = jwt.verify(token, process.env.SECRET_KEY);
+    return { id, username };
 };
 module.exports = { addIncome, getIncome, deleteIncome };
